+++ packages/data-stores/src/shared-types.ts
@@ -3,6 +3,16 @@ export interface WpcomClientCredentials {
 	client_secret: string;
 }
 
+export type FeatureId =
+	| 'domain'
+	| 'store'
+	| 'seo'
+	| 'plugins'
+	| 'ad-free'
+	| 'image-storage'
+	| 'video-storage'
+	| 'support';
+
 declare global {
 	interface Window {
 		_currentSiteId: number;
